refactor card UI & order structs
