+++ recipes/food-ordering/android/food-ordering/src/main/java/ai/picovoice/foodordering/Steps/PorcupineStep.java
@@ -30,7 +30,7 @@ public class PorcupineStep extends Step {
         super(r, listener);
         porcupine = new Porcupine.Builder()
                 .setAccessKey(accessKey)
-                .setKeywordPath(accessKey)
+                .setKeywordPath(modelPath)
                 .build(context);
     }
